# definition/api.yaml
openapi: 3.0.1
info:
  title: Boilerplate API
  version: 1.0.0
tags:
  - name: auth
    description: auth service
paths:
  /v1/auth/login:
    post:
      tags:
        - auth
      summary: Log in an existing user
      description: Log in an existing user
      operationId: Login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Success
          headers:
            X-Jwt:
              schema:
                type: string
              description: JWT Token to be used for authentication purpose
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjdmYmI1OWRhLWRiOTAtNGIwNi1iMjljLTFjZGZmYzQ4YTY5OCIsInBob25lTm8iOiIwODE5Njg4ODY5IiwiZmlyZWJhc2VUb2tlbiI6ImV5SmhiR2NpT2lKSVV6STFOaUlzSW5SNWNDSTZJa3BYVkNKOS5leUp6ZFdJaU9pSXhNak0wTlRZM09Ea3dJaXdpYm1GdFpTSTZJa3B2YUc0Z1JHOWxJaXdpYVdGMElqb3hOVEUyTWpNNU1ESXlmUS5TZmxLeHdSSlNNZUtLRjJRVDRmd3BNZUpmMzZQT2s2eUpWX2FkUXNzdzVjIiwicG9pbnRCYWxhbmNlIjoyMDAxOSwicmFuayI6Ikp1cmFnYW4iLCJxdWVzdHMiOlt7ImlkIjoiMzRiYmI3MjYtYjZhYS00ZDdjLTk3ZjUtMTAyMzUzZDc5ODBhIiwidHlwZSI6IlRBUkdFVCIsIm5hbWUiOiJKYW51YXJpIENlcmlhIiwiaW5zdHJ1Y3Rpb24iOiJLdW1wdWxrYW4gMi4wMDAgdG9uIHNlcGFuamFuZyBidWxhbiBKYW51YXJpIDIwMjMgdW50dWsgbWVuZGFwYXRrYW4gMjAuMDAwIHBvaW4iLCJleHBpcnlEYXRlIjoiMjAyMy0wMS0zMSAyMzo1OTo1OSIsImRhdGEiOlt7InJlcXVpcmVtZW50IjoyMDAwLCJwcm9ncmVzcyI6MjAwMCwidW5pdCI6InRvbiJ9XX0seyJpZCI6IjdiZTA0MDBmLWRkYjEtNGE5MS05OWQ5LTEzNjlkMDJmODg0NSIsInR5cGUiOiJXT1JLRVIiLCJuYW1lIjoiQWt0aXZpdGFzIFBla2VyamEiLCJpbnN0cnVjdGlvbiI6bnVsbCwiZXhwaXJ5RGF0ZSI6bnVsbCwiZGF0YSI6W3siaWQiOiI1YjU2ZjE2NS0wODVhLTRmMzEtYTgxNS0wZjRlZDJkY2FiZGMiLCJuYW1lIjoiU3VwYXJtYW4iLCJwaG9uZU5vIjoiMDgxMjEzMTQxNTE2Iiwiam9iIjoiS29vcmRpbmF0b3IgTGFwYW5nYW4iLCJwb2ludHMiOiIxIHBvaW4ifV19LHsiaWQiOiIzNWQxOWU2ZC1mMTEyLTQ1YTItYmVjOS00MmZjOGY4NzBmMTIiLCJ0eXBlIjoiU1VSVkVZIiwibmFtZSI6IlBhcnRpc2lwYXNpIFN1cnZleSIsImluc3RydWN0aW9uIjoiSXNpIHN1cnZleSBrZXB1YXNhbiBQYWJyaWsgQnVhdGFuIFNhdHUiLCJleHBpcnlEYXRlIjoiMjAyMy0wMS0zMSAyMzo1OTo1OSJ9XSwibGFzdFVwZGF0ZWQiOiIyMDIzLTAxLTMxIDIzOjU5OjU5IiwicmVsYXRlZE1pbGxzIjpbeyJpZCI6IjE0YjE5NjJjLTk3MGMtNDMwYi1hMjlhLTNhNzMwM2FhNTc1OSIsIm5hbWUiOiJQYWJyaWsgQnVhdGFuIFNhdHUifV0sImFzc2V0VHlwZXMiOlsiRE8iLCJUUlVDSyJdLCJsaW5rUmVmZXJyYWwiOiJodHRwczovL2FwaS5zYXdpdHByby5jb20vcmVmPTEyMzEyMzEyMzEyMzEyMzEyMzEyMzEyMyIsImlhdCI6MTY3NjM2NTM5NH0.6x5ic5p4BlqXhOA-dzDYD9Bev2_qMJuN_9jORsvEjTI
              required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
        '400':
          description: Bad Request - Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequestResponse'
        '401':
          description: Unauthorized - If JWT Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnAuthorizedResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
  /v1/auth/register:
    post:
      summary: Register a new user
      description: Register a new user
      operationId: Register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRegisterResponse'
        '400':
          description: Bad Request - Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequestResponse'
        '401':
          description: Unauthorized - If JWT Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnAuthorizedResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
  /v1/auth/verification:
    get:
      summary: Verify a user's email
      description: Verify a user's email
      operationId: Verification
      parameters:
        - $ref: '#/components/parameters/EmailParameter'
        - $ref: '#/components/parameters/VerificationTokenParameter'
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthVerificationResponse'
        '400':
          description: Bad Request - Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequestResponse'
        '401':
          description: Unauthorized - If JWT Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnAuthorizedResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
  /v1/auth/me:
    get:
      summary: Retrieve user information
      description: Retrieve user information
      operationId: Me
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthMeResponse'
        '400':
          description: Bad Request - Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequestResponse'
        '401':
          description: Unauthorized - If JWT Validation Failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnAuthorizedResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        Authorization using header parameter with parameter name `Authorization`
        See: https://swagger.io/docs/specification/authentication/bearer-authentication/
      bearerFormat: Bearer <token>
  parameters:
    EmailParameter:
      name: email
      in: query
      required: true
      description: User's email to verify
      schema:
        type: string
        format: email
        example: person@mail.com
    VerificationTokenParameter:
      name: verificationToken
      in: query
      required: true
      description: Verification token for the user
      schema:
        type: string
        example: 1234567890ASDOICASC
  schemas:
    AuthRegisterRequest:
      type: object
      required:
        - full_name
        - email
        - password
        - phone_number
      properties:
        full_name:
          type: string
        email:
          type: string
        password:
          type: string
        phone_number:
          type: string
    AuthRegisterResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    AuthLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
        password:
          type: string
    AuthLoginResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    AuthVerificationRequest:
      type: object
      required:
        - email
        - verification_token
      properties:
        email:
          type: string
        verification_token:
          type: string
    AuthVerificationResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    AuthMeResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ErrorBadRequestResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorMessage'
    ErrorUnAuthorizedResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Message explaining the error that happened
          example: 'Authorization failed, invalid credentials'
    ErrorInternalServerResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorMessage'
    ErrorMessage:
      type: string
      example: 'Error happened, please check service log'